--- src/apps/pflua_packet_filter/samples/COPYRIGHT
@@ -1,3 +1,0 @@
-v4.pcap and v6.pcap are from http://wiki.wireshark.org/SampleCaptures
-
-http://wiki.wireshark.org/License
